Remove YourImpact
