--- Client/Ecosia/UI/Impact/YourImpactLayout.swift
@@ -1,26 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import UIKit
-
-final class YourImpactLayout: UICollectionViewFlowLayout {
-    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
-        let attributes = super.layoutAttributesForElements(in: rect)?.map { $0.copy() } as? [UICollectionViewLayoutAttributes]
-        attributes?.reduce([CGFloat: (CGFloat, [UICollectionViewLayoutAttributes])]()) {
-            guard $1.representedElementCategory == .cell else { return $0 }
-            return $0.merging([ceil($1.center.y): ($1.frame.origin.y, [$1])]) {
-                ($0.0 < $1.0 ? $0.0 : $1.0, $0.1 + $1.1)
-            }
-        }
-        .values.forEach { minY, line in
-            line.forEach {
-                $0.frame = $0.frame.offsetBy(
-                    dx: 0,
-                    dy: minY - $0.frame.origin.y
-                )
-            }
-        }
-        return attributes
-    }
-}
