+++ Client/Ecosia/UI/NTP/AboutEcosia/NTPAboutEcosiaCell.swift
@@ -206,6 +206,7 @@ final class NTPAboutEcosiaCell: UICollectionViewCell, ReusableCell {
     @objc private func learnMoreAction() {
         guard let section = section else { return }
         viewModel?.delegate?.openLink(url: section.url)
+        Analytics.shared.navigation(.open, label: section.label) // TODO: Verify Analytics
     }
 }
