+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+WebViewDelegates.swift
@@ -587,13 +587,6 @@ extension BrowserViewController: WKNavigationDelegate {
             return
         }
 
-        // Ecosia: intercept URL to impact
-        if Environment.current.urlProvider.isYourImpact(url: url),
-           navigationAction.navigationType == .linkActivated {
-            presentYourImpact()
-            decisionHandler(.cancel)
-        }
-
         // This is the normal case, opening a http or https url, which we handle by loading them in this WKWebView. We
         // always allow this. Additionally, data URIs are also handled just like normal web pages.
